Handle browser running/not running in Start and Stop
If no browser is running we don't have to do anything in Stop. Contrary, if a
browser is already running we might be confused if starting it again would just
re-use that browser (unexpected config) so it seems safer to error there.

+++ headless.go
@@ -106,6 +106,9 @@ func (b *Browser) Start(url string) error {
 }
 
 func (b *Browser) Stop() error {
+	if b.cmd == nil {
+		return nil
+	}
 	cmd := b.cmd
 	b.cmd = nil
 	if err := cmd.Process.Kill(); err != nil {
@@ -119,6 +122,9 @@ func (b *Browser) Stop() error {
 }
 
 func (h *H) Start() error {
+	if h.Browser.cmd != nil {
+		return fmt.Errorf("browser already running")
+	}
 	if h.Port == 0 {
 		h.Port = GetFreePort()
 	}
